Refactor
Move User Group and Role class to principal.py
