+++ fusillade/__init__.py
@@ -1,5 +1,5 @@
 from .config import Config
 from . import logging
-from .directory import User, Group, Role
+from fusillade.directory.principal import User, Group, Role
 from fusillade.directory.clouddirectory import CloudDirectory
 from .errors import FusilladeException, FusilladeBindingException
